+++ app/src/main/java/com/example/tk/activity/UserMainActivity.java
@@ -1,4 +1,4 @@
-package com.example.tk;
+package com.example.tk.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
@@ -9,6 +9,11 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.Toast;
+
+import com.example.tk.R;
+import com.example.tk.backActivity.Insertuser_Activity;
+import com.example.tk.backActivity.Sea_deluser_Activity;
+import com.example.tk.userDatabase.user_database;
 
 public class UserMainActivity extends AppCompatActivity {
     public user_database user;
