+++ app/src/main/java/com/example/tk/activity/UserMainMeActivity.java
@@ -1,4 +1,4 @@
-package com.example.tk;
+package com.example.tk.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,18 +7,21 @@ import android.widget.LinearLayout;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class UserMainMessageActivity extends AppCompatActivity{
+import com.example.tk.SignLog.LoginActivity;
+import com.example.tk.R;
+
+public class UserMainMeActivity extends AppCompatActivity{
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.main_message_layout);
+        setContentView(R.layout.main_me_layout);
 
-        LinearLayout messageView = findViewById(R.id.linearLayout1);
+        LinearLayout messageView = findViewById(R.id.linearLayout2);
         messageView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(UserMainMessageActivity.this, UserMainActivity.class);
+                Intent intent = new Intent(UserMainMeActivity.this, UserMainMessageActivity.class);
                 startActivity(intent);
                 overridePendingTransition(0, 0);
                 finish();
@@ -29,20 +32,29 @@ public class UserMainMessageActivity extends AppCompatActivity{
         newView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(UserMainMessageActivity.this, UserMainNewActivity.class);
+                Intent intent = new Intent(UserMainMeActivity.this, UserMainNewActivity.class);
                 startActivity(intent);
                 overridePendingTransition(0, 0);
                 finish();
             }
         });
-        LinearLayout meView = findViewById(R.id.linearLayout4);
+        LinearLayout meView = findViewById(R.id.linearLayout1);
         meView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(UserMainMessageActivity.this, UserMainMeActivity.class);
+                Intent intent = new Intent(UserMainMeActivity.this, UserMainActivity.class);
                 startActivity(intent);
                 overridePendingTransition(0, 0);
                 finish();
+            }
+        });
+
+        LinearLayout loginView = findViewById(R.id.login);
+        loginView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(UserMainMeActivity.this, LoginActivity.class);
+                startActivity(intent);
             }
         });
     }
