+++ app/src/main/java/com/example/tk/dao/UserInfo.java
@@ -1,4 +1,4 @@
-package com.example.tk;
+package com.example.tk.dao;
 
 public class UserInfo {
     private int id;
